==> 2. Extract Dublin Weather Forecast.ipynb <==
## <u>Notebook by John Uzoma</u>

#### Get XML data and write to bronze layer

import requests
import ssl
import os

# Enable SSL certificate verification warnings
ctx = ssl.create_default_context()
ctx.check_hostname = True
ctx.verify_mode = ssl.CERT_REQUIRED

# Define the URL from which to fetch the XML data
url = "https://metwdb-openaccess.ichec.ie/metno-wdb2ts/locationforecast?lat=53.35014;long=-6.266155"

# Send a GET request to the URL and fetch the XML data
response = requests.get(url, verify=True)
xml_data = response.content

# Convert bytes data to string
xml_data = xml_data.decode('utf-8')

# Convert the string to a list of tuples (each tuple contains one string)
data = [(xml_data,)]

# Define the output file path
output_file_path = "Files/Bronze/DublinWeatherForecast.txt"

# Create a DataFrame from the list of tuples
df = spark.createDataFrame(data, ["text"])

# Write the DataFrame to a text file
df.write.mode("overwrite").text(output_file_path)
